requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
python-docx>=0.8.11
openpyxl>=3.1.0
